User renamed into UserAccount
The User is associated to the user who owns the resources.
The UserAccount is one of the accounts the User has.

Both User and UserAccount will be propagated in the access token. This
will allow to work as before, but some extensions such as the Id Token
Grant Token or the UserInfo extension will be able to support multpile
identities of that User